feat(Symfony): boot messaging system on service injection
