+++ packages/Symfony/SymfonyBundle/EcotoneSymfonyBundle.php
@@ -30,44 +30,6 @@ class EcotoneSymfonyBundle extends Bundle
     public function build(ContainerBuilder $container)
     {
         $container->addCompilerPass(new EcotoneCompilerPass());
-
-        $this->setUpExpressionLanguage($container);
-
-        $definition = new Definition();
-        $definition->setClass(ConfiguredMessagingSystem::class);
-        $definition->setSynthetic(true);
-        $definition->setPublic(true);
-        $container->setDefinition(ConfiguredMessagingSystem::class, $definition);
-    }
-
-    /**
-     * @param ContainerBuilder $container
-     * @return void
-     */
-    private function setUpExpressionLanguage(ContainerBuilder $container): void
-    {
-        $expressionLanguageAdapter = ExpressionEvaluationService::REFERENCE . '_adapter';
-        $definition = new Definition();
-        $definition->setClass(ExpressionLanguage::class);
-
-        $container->setDefinition($expressionLanguageAdapter, $definition);
-
-        $definition = new Definition();
-        $definition->setClass(SymfonyExpressionEvaluationAdapter::class);
-        $definition->setFactory([SymfonyExpressionEvaluationAdapter::class, 'create']);
-        $definition->setPublic(true);
-        $container->setDefinition(ExpressionEvaluationService::REFERENCE, $definition);
-    }
-
-    public function boot()
-    {
-        $configuration = EcotoneCompilerPass::getMessagingConfiguration($this->container, true);
-
-        $referenceSearchService = $this->container->get(ReferenceSearchService::class);
-        $messagingSystem = $configuration->buildMessagingSystemFromConfiguration($referenceSearchService);
-        $referenceSearchService->setConfiguredMessagingSystem($messagingSystem);
-
-        $this->container->set(ConfiguredMessagingSystem::class, $messagingSystem);
     }
 
     public function getContainerExtension(): ?ExtensionInterface
