+++ packages/Symfony/composer.json
@@ -37,6 +37,7 @@
         "phpstan/phpstan": "^1.8",
         "phpunit/phpunit": "^9.5",
         "symfony/doctrine-messenger": "^5.4|^6.0",
+        "symfony/expression-language": "^6.0",
         "symfony/messenger": "^5.4|^6.0",
         "wikimedia/composer-merge-plugin": "^2.0"
     },
